Refactor: Email sender functions into a class helper
